requirements_integration.txt
# Requirements for PLC -> OPC UA -> MES -> DATABASE Integration
# EV Manufacturing System Dependencies

# Core Python packages
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# OPC UA Communication
asyncua==1.0.7
cryptography>=41.0.0

# Database and Data Processing
sqlite3  # Built into Python
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Web Framework and APIs
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio[client]>=5.8.0

# MQTT Communication
paho-mqtt>=1.6.1
python-socketio>=5.8.0

# Data Serialization and Validation
pydantic>=2.0.0
marshmallow>=3.20.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Time and Date Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Mathematical and Scientific Computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Networking and HTTP
requests>=2.31.0
httpx>=0.24.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Visualization and Plotting
matplotlib>=3.7.0
plotly>=5.15.0
dash>=2.12.0

# System Monitoring
psutil>=5.9.0
watchdog>=3.0.0

# Data Export and Reporting
openpyxl>=3.1.0
jinja2>=3.1.0

# Performance and Optimization
uvloop>=0.17.0  # For Linux/macOS
cachetools>=5.3.0

# Security
bcrypt>=4.0.0
jwt>=1.3.1

# Optional PLC Communication Libraries
# Uncomment based on your PLC type:
# pymodbus>=3.5.0          # For Modbus communication
# snap7>=1.3               # For Siemens S7 PLCs
# pycomm3>=1.2.0           # For Allen-Bradley PLCs
# pyads>=3.3.0             # For Beckhoff TwinCAT ADS

# Optional Industrial Protocols
# python-ethercat>=1.0.0   # For EtherCAT
# canopen>=2.2.0           # For CANopen
# profibus>=1.0.0          # For PROFIBUS
